Refactor exceptions into packages
