+++ src/main/java/seedu/duke/commands/food/AddFoodCommand.java
@@ -3,7 +3,7 @@ package seedu.duke.commands.food;
 import seedu.duke.commands.Command;
 import seedu.duke.exceptions.ArgumentsNotFoundException;
 import seedu.duke.food.FoodRecord;
-import seedu.duke.exceptions.IllegalFoodParameterException;
+import seedu.duke.exceptions.food.IllegalFoodParameterException;
 import seedu.duke.parser.Parser;
 import seedu.duke.storage.Storage;
 import seedu.duke.storage.StorageFood;
