+++ src/main/java/seedu/duke/commands/journal/AddEntryCommand.java
@@ -3,9 +3,9 @@ package seedu.duke.commands.journal;
 import seedu.duke.commands.Command;
 import seedu.duke.exceptions.EmptyEntryArgumentsException;
 import seedu.duke.exceptions.EmptyEntryNameException;
-import seedu.duke.exceptions.EmptyNoteNameException;
-import seedu.duke.exceptions.NotebookArgumentNotFoundException;
-import seedu.duke.exceptions.NotebookNotFoundForEntryAddition;
+import seedu.duke.exceptions.journal.EmptyNoteNameException;
+import seedu.duke.exceptions.journal.NotebookArgumentNotFoundException;
+import seedu.duke.exceptions.journal.NotebookNotFoundForEntryAddition;
 import seedu.duke.journal.Note;
 import seedu.duke.parser.journal.ParserJournal;
 import seedu.duke.storage.Storage;
