+++ src/main/java/seedu/duke/commands/journal/AddNoteCommand.java
@@ -1,10 +1,10 @@
 package seedu.duke.commands.journal;
 
 import seedu.duke.commands.Command;
-import seedu.duke.exceptions.DuplicateNoteException;
-import seedu.duke.exceptions.EmptyNoteArgumentsException;
-import seedu.duke.exceptions.EmptyNoteNameException;
-import seedu.duke.exceptions.NotebookArgumentNotFoundException;
+import seedu.duke.exceptions.journal.DuplicateNoteException;
+import seedu.duke.exceptions.journal.EmptyNoteArgumentsException;
+import seedu.duke.exceptions.journal.EmptyNoteNameException;
+import seedu.duke.exceptions.journal.NotebookArgumentNotFoundException;
 import seedu.duke.parser.journal.ParserJournal;
 import seedu.duke.storage.Storage;
 import seedu.duke.storage.StorageNotes;
