+++ src/main/java/seedu/duke/commands/module/AddModuleCommand.java
@@ -2,7 +2,7 @@ package seedu.duke.commands.module;
 
 import seedu.duke.commands.Command;
 import seedu.duke.exceptions.ClickException;
-import seedu.duke.exceptions.IllegalModuleException;
+import seedu.duke.exceptions.module.IllegalModuleException;
 import seedu.duke.module.Module;
 import seedu.duke.module.ModuleList;
 import seedu.duke.storage.Storage;
