--- src/main/java/seedu/duke/exceptions/IllegalModuleException.java
@@ -1,4 +1,0 @@
-package seedu.duke.exceptions;
-
-public class IllegalModuleException extends ClickException {
-}
