--- src/main/java/seedu/duke/exceptions/IllegalModuleIndexException.java
@@ -1,4 +1,0 @@
-package seedu.duke.exceptions;
-
-public class IllegalModuleIndexException extends ClickException {
-}
