+++ src/main/java/seedu/duke/exceptions/journal/DuplicateNoteException.java
@@ -1,4 +1,4 @@
-package seedu.duke.exceptions;
+package seedu.duke.exceptions.journal;
 
 public class DuplicateNoteException extends Exception {
     public DuplicateNoteException() {
