+++ src/main/java/seedu/duke/exceptions/journal/IncorrectJournalArgumentException.java
@@ -1,4 +1,4 @@
-package seedu.duke.exceptions;
+package seedu.duke.exceptions.journal;
 
 public class IncorrectJournalArgumentException extends Exception {
     public IncorrectJournalArgumentException() {
