+++ src/main/java/seedu/duke/exceptions/journal/NotebookArgumentNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.duke.exceptions;
+package seedu.duke.exceptions.journal;
 
 public class NotebookArgumentNotFoundException extends Exception {
     public NotebookArgumentNotFoundException(String message) {
