+++ src/main/java/seedu/duke/exceptions/journal/NotebookNotFoundForEntryAddition.java
@@ -1,4 +1,4 @@
-package seedu.duke.exceptions;
+package seedu.duke.exceptions.journal;
 
 public class NotebookNotFoundForEntryAddition extends Exception {
     public NotebookNotFoundForEntryAddition() {
